+++ lib/hubbado/log/notify_rollbar.rb
@@ -17,7 +17,7 @@ module Hubbado
         @notifier = notifier
       end
 
-      def log(subject, severity, message, data = nil, stacktrace = nil)
+      def log(subject, severity, message, data = nil, stacktrace = nil, _tags = [])
         level = LEVELS[severity.to_sym]
         return if level.nil?