+++ lib/hubbado/log/rails_logger.rb
@@ -14,7 +14,9 @@ module Hubbado
         @rails_logger = rails_logger
       end
 
-      def log(subject, severity, message, data = nil, stacktrace = nil)
+      def log(subject, severity, message, data = nil, stacktrace = nil, tags = [])
+        return unless Display.shows?(severity, tags)
+
         rails_severity = RAILS_SEVERITIES.fetch(severity.to_sym, severity)
 
         rails_logger.send(rails_severity, "#{subject}: #{message}")